requirements.txt
# Core system requirements for Autonomous Multi-Agent System
pytest>=7.0.0
pytest-asyncio>=0.21.0

# LLM Integration - UPDATED
aiohttp>=3.8.0
openai>=1.0.0  # For OpenAI API integration
google-generativeai>=0.3.0  # For Gemini API integration
python-dotenv>=1.0.0  # For environment variables from .env file

# Token counting for document analysis
tiktoken>=0.5.0

# Async file operations
aiofiles>=23.0.0
aiosqlite>=0.21.0

# Security scanning
bandit>=1.7.0
safety>=2.0.0

# Code quality
flake8>=6.0.0
coverage>=7.0.0

# Machine Learning and Data Science
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Knowledge Graph and Graph Database
networkx>=3.0
py2neo>=2021.2.4
neo4j>=5.0.0

# Additional Python packages that may be needed
requests>=2.25.0
pyyaml>=6.0

# Web framework for Flask app
flask>=2.0.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=1.4.0

# Testing framework extensions
pytest-flask>=1.3.0
